# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API proyecto EVA",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/evaluacion": {
            "get": {
                "tags": [
                    "Evaluaciones"
                ],
                "summary": "Obtiene una lista de los evaluacions",
                "operationId": "7881e5c31a0a53bb3eb148548da6f57d",
                "responses": {
                    "200": {
                        "description": "Lista de evaluacions"
                    }
                }
            },
            "post": {
                "tags": [
                    "Evaluaciones"
                ],
                "summary": "Crear un nuevo evaluacion",
                "operationId": "78d1e532f228132082d857405a938612",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "fecha_revision",
                                    "hora_revision",
                                    "concluido",
                                    "nota",
                                    "id_empresa",
                                    "id_tipo_evaluacion"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Item 1"
                                    },
                                    "fecha_revision": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024/10/21"
                                    },
                                    "hora_revision": {
                                        "type": "time",
                                        "example": "18:00"
                                    },
                                    "nota": {
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "id_empresa": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "id_tipo_evaluacion": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "evaluacion creada con éxito"
                    },
                    "422": {
                        "description": "Problemas con los datos ingresados"
                    }
                }
            }
        },
        "/api/evaluacion/{id}": {
            "get": {
                "tags": [
                    "Evaluaciones"
                ],
                "summary": "Mostar un evaluacion",
                "operationId": "1a06bbcda976cfa88c85b1ffda1e7b29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la evaluacion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la evaluacion"
                    },
                    "404": {
                        "description": "evaluacion no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Evaluaciones"
                ],
                "summary": "Actualizar un evaluacion",
                "operationId": "ab96158ade3c36b79cf6bcfb7769aa5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la evaluacion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Item 1"
                                    },
                                    "fecha_revision": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024/10/21"
                                    },
                                    "hora_revision": {
                                        "type": "time",
                                        "example": "18:00"
                                    },
                                    "nota": {
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "id_empresa": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "id_tipo_evaluacion": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "evaluacion actualizada con éxito"
                    },
                    "404": {
                        "description": "evaluacion no encontrada"
                    },
                    "422": {
                        "description": "Problemas con los datos ingresados"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Evaluaciones"
                ],
                "summary": "Eliminar una evaluacion",
                "operationId": "5ef3dd512fe70fb2572123ce749f15d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la evaluacion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "evaluacion eliminada "
                    },
                    "404": {
                        "description": "evaluacion no encontrada"
                    }
                }
            }
        },
        "/api/evaluacion/empresa/{id}": {
            "delete": {
                "tags": [
                    "Evaluaciones"
                ],
                "summary": "Mostar una evaluaciones por empresa",
                "operationId": "392328bc1513e1d0fc998815ae24f3b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "muestra evaluacion "
                    },
                    "404": {
                        "description": "evaluacion no encontrada"
                    }
                }
            }
        },
...
        "/api/planilla_seguimiento/semanal/{idUsuario}": {
            "get": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Mostar planillas Seguimientos por docente",
                "operationId": "bfe745f507e29ceee65e862596bc018e",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "description": "ID de el docente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del planilla Seguimiento por docente"
                    },
                    "404": {
                        "description": "docente no encontrada"
                    }
                }
            }
        },
...
    },
...
}